add admin and tokens and password hashed

+++ main.py
@@ -1,28 +1,53 @@
 from fastapi import FastAPI, Depends, HTTPException
 from database import create_db_and_tables ,get_session 
-from models import Accounts, Transactions
+from models import Accounts, Transactions,UserRole
 from sqlmodel import Session, select
-from schemas import AccountAdd,AccountRead,BalanceDeposi,BalanceWithdraw,TransactionRead
-import security
+from schemas import AccountUser,AccountRead,BalanceDeposi,BalanceWithdraw,TransactionRead,LoginUser,AccountAdmin
+from security import hash_password, verify_password, create_access_token, get_current_user, get_current_admin
+from admin_setup import create_main_admin, admin_num
+
+
 
 
 app = FastAPI()
 
+
+
 @app.on_event("startup")
 def startup_event():
     create_db_and_tables()
+    create_main_admin()
+
+
     
 
-@app.get("/")
-def the_welcome_space():
-    return {"version":"0.0.0"}
+@app.post("/login")
+def login(data: LoginUser, session: Session = Depends(get_session)):
+    email = data.email
+    password = data.password
+    user = session.exec(select(Accounts).where(Accounts.email == email)).first()
+        
+    if not user:
+        raise HTTPException(status_code=400, detail="invalid email")
+    
+    if not verify_password(password, user.hashed_password):
+        raise HTTPException(status_code=400, detail= "invalid password")  
+        
+    access_token = create_access_token(
+        data={"sub": user.email}
+    ) 
+    return{
+        "access_token": access_token,
+        "token_type": "bearer"
+    } 
+          
 
-@app.post("/create_accounts")
-def creat_account(account:AccountAdd,
+@app.post("/register")
+def creat_user(account:AccountUser,
                   session: Session = Depends(get_session)
     ):
     
-    hashed_pw = security.hash_password(account.password)    
+    hashed_pw = hash_password(account.password)    
     db_account = Accounts(
         user_name = account.user_name,
         email = account.email,
@@ -41,11 +66,38 @@ def creat_account(account:AccountAdd,
     return db_account
 
 
+@app.post("/creat_admin")
+def creat_admin(
+    admin_data: AccountAdmin,
+    session: Session = Depends(get_session),
+    current_admin: Accounts = Depends(get_current_admin)
+):
+    existing_user = session.exec(
+        select(Accounts).where(Accounts.email == admin_data.email)
+    ).first()
+    
+    if existing_user:
+        raise HTTPException(status_code=400, detail="Email already exists")
+    
+    new_admin = Accounts(
+        user_name=admin_data.user_name,
+        email=admin_data.email,
+        compte_num=admin_num(),
+        hashed_password=hash_password(admin_data.password),
+        role=UserRole.admin
+    )
+    session.add(new_admin)
+    session.commit()
 
-@app.put("/deposit/{user_id}", response_model= BalanceDeposi)
-def deposit_money(user_id: int, new_balance:BalanceDeposi ,session : Session = Depends(get_session)): 
+    return {"message": "Admin created successfully"}
+
+@app.put("/deposit", response_model= BalanceDeposi)
+def deposit_money(new_balance:BalanceDeposi ,
+                  session : Session = Depends(get_session),
+                  current_user: str = Depends(get_current_user)
+                  ): 
             
-    account = session.exec(select(Accounts).where(Accounts.id == user_id)).first()
+    account = session.exec(select(Accounts).where(Accounts.email == current_user )).first()
     
     if not account:
         raise HTTPException(status_code=404, detail="Account not found")
@@ -56,7 +108,7 @@ def deposit_money(user_id: int, new_balance:BalanceDeposi ,session : Session = D
     amount = account.deposit(new_balance.deposit)
     
     
-    new_transaction = Transactions(account_id = user_id,
+    new_transaction = Transactions(account_id = account.id,
                                    type = "deposit",
                                    amount = new_balance.deposit
                                    )
@@ -67,10 +119,13 @@ def deposit_money(user_id: int, new_balance:BalanceDeposi ,session : Session = D
     return {"deposit" : amount}
 
 
-@app.put("/withdraw/{user_id}", response_model = BalanceWithdraw)
-def withdraw_money(user_id: int, new_balance:BalanceWithdraw ,session : Session = Depends(get_session)): 
+@app.put("/withdraw", response_model = BalanceWithdraw)
+def withdraw_money( new_balance:BalanceWithdraw ,
+                   session : Session = Depends(get_session),
+                   current_user: str = Depends(get_current_user)
+                   ): 
     
-    account = session.exec(select(Accounts).where(Accounts.id == user_id)).first()
+    account = session.exec(select(Accounts).where(Accounts.email == current_user)).first()
     
     if not account:
         raise HTTPException(status_code=404, detail="Account not found")
@@ -80,7 +135,7 @@ def withdraw_money(user_id: int, new_balance:BalanceWithdraw ,session : Session 
     
     amount = account.withdraw(new_balance.withdraw)
     
-    new_transaction = Transactions(account_id= user_id,
+    new_transaction = Transactions(account_id= account.id,
                                    type= "withdraw",
                                    amount = new_balance.withdraw
                                    )
@@ -92,7 +147,10 @@ def withdraw_money(user_id: int, new_balance:BalanceWithdraw ,session : Session 
     
     
 @app.get("/accounts", response_model=list[AccountRead])
-def show_accounts(session: Session = Depends(get_session)):
+def show_accounts(
+    session: Session = Depends(get_session),
+    admin: Accounts = Depends(get_current_admin)
+    ):
     accounts = session.exec(select(Accounts)).all()
     return [AccountRead.from_orm(a) for a in accounts]  # ✅ Convert each
 
@@ -116,13 +174,19 @@ def transaction(user_id: int, session : Session = Depends(get_session)):
     
 
 @app.delete("/delete/{user_id}")
-def delete_account(user_id: int, session : Session = Depends(get_session)):
+def delete_account(
+    user_id: int,
+    session : Session = Depends(get_session),
+    cerent_admin: Accounts = Depends(get_current_admin)
+    ):
     account = session.exec(select(Accounts).where(Accounts.id == user_id)).first()
     if not account:
-         raise HTTPException(status_code=404, detail="Account not found")
+        raise HTTPException(status_code=404, detail="Account not found")
+    
+    if account.id == cerent_admin.id:
+        raise HTTPException(status_code=400, detail="Admin cannot delete themselves")
+
     
     session.delete(account)
     session.commit()
     return {"status": "Account deleted"}
-
-
